requirements.txt
pymongo
python-dotenv
langchain
openai
streamlit
